requirements.txt
pyyaml>=6.0
pandas>=2.0.0
numpy>=1.24.0
librosa>=0.10.0
soundfile>=0.12.0
seaborn>=0.12.0
matplotlib>=3.7.0
scipy>=1.10.0
